refactoring

+++ neuralnetwork/module.py
@@ -3,8 +3,8 @@ class Module:
     def __init__(self):
         pass
 
-    def forward(self, input_features: int):
-      pass
+    def forward(self, input_features: float):
+        pass
 
     def backpropagation(self, output_error: float, learning_rate: float):
         pass
